+++ MetaMorpheus/GuiFunctions/MetaDraw/BioPolymerCoverage/BioPolymerGroupViewModel.cs
@@ -2,7 +2,6 @@
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Linq;
-using Readers;
 
 namespace GuiFunctions;
 
@@ -69,16 +68,3 @@ public class BioPolymerGroupViewModel : BaseViewModel
         OnPropertyChanged(nameof(MaximumSequenceCoverage));
     }
 }
-
-public class BioPolymerCoverageResultModel(SpectrumMatchFromTsv match, string baseSequence, int start, int end, BioPolymerCoverageType coverageType)
-{
-    public readonly SpectrumMatchFromTsv Match = match;
-    public string BaseSequence { get; } = baseSequence;
-    public int Start { get; } = start;
-    public int End { get; } = end;
-    public BioPolymerCoverageType CoverageType { get; } = coverageType;
-    public double QValue => Match.QValue;
-    public string Ambiguity => Match.AmbiguityLevel;
-    public string FileName => Match.FileName;
-}
-
